refactor: wave 2 — extract shared, scheduled, whatsapp, metrics, fee-agreements, admin modules + remove dead code (approve/reject deprecated, 4 dead frontend files)
