+++ functions/index.js
@@ -7,11 +7,12 @@
 
 const functions = require('firebase-functions');
 const admin = require('firebase-admin');
-const { onSchedule } = require('firebase-functions/v2/scheduler');
-const { onDocumentWritten } = require('firebase-functions/v2/firestore');
 const { onRequest } = require('firebase-functions/v2/https');
 const { addTimeToTaskWithTransaction } = require('./addTimeToTask_v2');
 const { updateBudgetTask, markNotificationAsRead } = require('./task-update-realtime');
+const { checkUserPermissions } = require('./shared/auth');
+const { logAction } = require('./shared/audit');
+const { sanitizeString, isValidIsraeliPhone, isValidEmail } = require('./shared/validators');
 
 // ✨ NEW: Import modular deduction system
 const DeductionSystem = require('./src/modules/deduction');
@@ -23,11 +24,6 @@ const { generateCaseNumberWithTransaction } = require('./case-number-transaction
 admin.initializeApp();
 const db = admin.firestore();
 const auth = admin.auth();
-
-// Twilio environment variables for v1 functions compatibility
-const TWILIO_ACCOUNT_SID = process.env.TWILIO_ACCOUNT_SID || 'AC9e5e9e3c953a5bbb878622b6e70201b6';
-const TWILIO_AUTH_TOKEN = process.env.TWILIO_AUTH_TOKEN || 'fed2170530e4ed34d3b1b3407e0f0f5f';
-const TWILIO_WHATSAPP_NUMBER = process.env.TWILIO_WHATSAPP_NUMBER || 'whatsapp:+14155238886';
 
 // ===============================
 // CORS Configuration
@@ -42,108 +38,6 @@ const corsHeaders = {
 // ===============================
 // Helper Functions - פונקציות עזר
 // ===============================
-
-/**
- * בדיקת הרשאות משתמש
- */
-async function checkUserPermissions(context) {
-  // בדיקה שהמשתמש מחובר
-  if (!context.auth) {
-    throw new functions.https.HttpsError(
-      'unauthenticated',
-      'נדרשת התחברות למערכת'
-    );
-  }
-
-  const uid = context.auth.uid;
-
-  // טעינת פרטי העובד מ-Firestore
-  const employeeSnapshot = await db.collection('employees')
-    .where('authUID', '==', uid)
-    .limit(1)
-    .get();
-
-  if (employeeSnapshot.empty) {
-    throw new functions.https.HttpsError(
-      'not-found',
-      'עובד לא נמצא במערכת'
-    );
-  }
-
-  const employeeDoc = employeeSnapshot.docs[0];
-  const employee = employeeDoc.data();
-
-  // בדיקה שהעובד פעיל
-  if (!employee.isActive) {
-    throw new functions.https.HttpsError(
-      'permission-denied',
-      'חשבון המשתמש לא פעיל. אנא פנה למנהל המערכת.'
-    );
-  }
-
-  return {
-    uid,
-    email: employeeDoc.id, // Document ID is EMAIL (industry standard)
-    username: employee.username, // Username for display only
-    employee: employee,
-    role: employee.role || 'employee'
-  };
-}
-
-/**
- * רישום לוג ביקורת (Audit Log)
- */
-async function logAction(action, userId, username, details = {}) {
-  try {
-    await db.collection('audit_log').add({
-      action,
-      userId,
-      username,
-      details,
-      timestamp: admin.firestore.FieldValue.serverTimestamp(),
-      userAgent: details.userAgent || null,
-      ipAddress: details.ipAddress || null
-    });
-  } catch (error) {
-    console.error('Failed to write audit log:', error);
-    // לא נזרוק שגיאה - logging לא צריך לעצור את הפעולה
-  }
-}
-
-/**
- * ניקוי HTML (מניעת XSS)
- *
- * ✅ Fixed: רק < ו-> מוחלפים (סיכון XSS אמיתי)
- * ✅ גרשיים (" ו-') ו-/ לא מוחלפים - שמירת data integrity
- *
- * Note: Frontend צריך להשתמש ב-safeText() או textContent בdisplay
- */
-function sanitizeString(str) {
-  if (typeof str !== 'string') return str;
-  return str
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;');
-    // Removed: .replace(/"/g, '&quot;') - causes data corruption
-    // Removed: .replace(/'/g, '&#x27;') - causes data corruption
-    // Removed: .replace(/\//g, '&#x2F;') - not an XSS risk
-}
-
-/**
- * אימות מספר טלפון ישראלי
- */
-function isValidIsraeliPhone(phone) {
-  if (!phone) return true; // אופציונלי
-  const cleanPhone = phone.replace(/[-\s]/g, '');
-  return /^0(5[0-9]|[2-4]|[7-9])\d{7}$/.test(cleanPhone);
-}
-
-/**
- * אימות אימייל
- */
-function isValidEmail(email) {
-  if (!email) return true; // אופציונלי
-  return /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(email);
-}
 
 /**
  * יצירה או קבלת תיק פנימי לעובד (Lazy Creation)
@@ -5449,71 +5343,6 @@ exports.linkAuthToEmployee = functions.https.onCall(async (data, context) => {
   }
 });
 
-// ===============================
-// Activity Logging & User Tracking
-// ===============================
-
-/**
- * רישום פעילות משתמש (Activity Log)
- * נקרא מ-activity-logger.js
- */
-exports.logActivity = functions.https.onCall(async (data, context) => {
-  try {
-    const user = await checkUserPermissions(context);
-
-    // Validation
-    if (!data.type || typeof data.type !== 'string') {
-      throw new functions.https.HttpsError(
-        'invalid-argument',
-        'חסר סוג פעילות'
-      );
-    }
-
-    if (!data.action || typeof data.action !== 'string') {
-      throw new functions.https.HttpsError(
-        'invalid-argument',
-        'חסר תיאור פעולה'
-      );
-    }
-
-    // רישום הפעילות
-    const activityData = {
-      type: sanitizeString(data.type),
-      action: sanitizeString(data.action),
-      details: data.details ? sanitizeString(JSON.stringify(data.details)) : '',
-      userId: user.uid,
-      username: user.username,
-      timestamp: admin.firestore.FieldValue.serverTimestamp(),
-      userAgent: data.userAgent || null,
-      sessionId: data.sessionId || null
-    };
-
-    const docRef = await db.collection('activity_log').add(activityData);
-
-    return {
-      success: true,
-      activityId: docRef.id
-    };
-
-  } catch (error) {
-    console.error('Error in logActivity:', error);
-
-    if (error instanceof functions.https.HttpsError) {
-      throw error;
-    }
-
-    throw new functions.https.HttpsError(
-      'internal',
-      `שגיאה ברישום פעילות: ${error.message}`
-    );
-  }
-});
-
-// ✅ trackUserActivity REMOVED - replaced by Firebase Realtime Database Presence
-// Old heartbeat-based tracking consumed 2,880 writes/day
-// New presence system uses only ~60 writes/day (98% reduction!)
-// See: js/modules/presence-system.js
-
 /**
  * setAdminClaim - מגדיר Custom Claim של admin למשתמש
  * מאפשר הרשאות מתקדמות ב-Security Rules
@@ -5661,475 +5490,9 @@ exports.initializeAdminClaims = functions.https.onCall(async (data, context) => 
   }
 });
 
-// ===============================
-// Scheduled Functions - פונקציות מתוזמנות
-// ===============================
-
-/**
- * dailyTaskReminders - תזכורות משימות יומיות
- * רץ כל יום בשעה 09:00 בבוקר
- * בודק:
- * 1. משימות שעומדות לפוג בתוך 3 ימים
- * 2. משימות שכבר עבר תאריך היעד שלהן (overdue)
- * שולח התראה אוטומטית לעובדים (לא למנהלים - הם רואים בדשבורד)
- */
-exports.dailyTaskReminders = onSchedule({
-  schedule: '0 9 * * *',  // כל יום בשעה 09:00
-  timeZone: 'Asia/Jerusalem',
-  region: 'us-central1'
-}, async (event) => {
-    try {
-      console.log('🔔 Running dailyTaskReminders at', new Date().toISOString());
-
-      const now = admin.firestore.Timestamp.now();
-      const threeDaysFromNow = new Date();
-      threeDaysFromNow.setDate(threeDaysFromNow.getDate() + 3);
-      const threeDaysTimestamp = admin.firestore.Timestamp.fromDate(threeDaysFromNow);
-
-      // מצא משימות פעילות עם deadline בתוך 3 ימים או שעבר
-      const tasksSnapshot = await db.collection('budget_tasks')
-        .where('status', '==', 'פעיל')
-        .where('deadline', '!=', null)
-        .get();
-
-      let remindersCount = 0;
-      let overdueCount = 0;
-
-      for (const taskDoc of tasksSnapshot.docs) {
-        const task = taskDoc.data();
-        const taskId = taskDoc.id;
-        const deadline = task.deadline;
-
-        // דלג על משימות ללא deadline
-        if (!deadline) continue;
-
-        const isOverdue = deadline.toDate() < now.toDate();
-        const isUpcoming = !isOverdue && deadline.toDate() <= threeDaysTimestamp.toDate();
-
-        if (isOverdue) {
-          // משימה שעבר הזמן
-          await db.collection('notifications').add({
-            userId: task.lawyer || task.createdBy,
-            userEmail: task.employee,
-            title: `⚠️ משימה באיחור: ${task.clientName}`,
-            message: `המשימה "${task.description}" עברה את תאריך היעד (${formatDate(deadline.toDate())})`,
-            type: 'error',
-            taskId: taskId,
-            reminder: true,
-            automated: true,
-            read: false,
-            createdAt: admin.firestore.FieldValue.serverTimestamp(),
-            icon: 'fa-exclamation-triangle'
-          });
-          overdueCount++;
-
-        } else if (isUpcoming) {
-          // משימה שמתקרבת לתאריך יעד
-          const daysLeft = Math.ceil((deadline.toDate() - now.toDate()) / (1000 * 60 * 60 * 24));
-          await db.collection('notifications').add({
-            userId: task.lawyer || task.createdBy,
-            userEmail: task.employee,
-            title: `⏰ תזכורת: ${task.clientName}`,
-            message: `המשימה "${task.description}" מתקרבת לתאריך יעד (${daysLeft} ימים)`,
-            type: 'warning',
-            taskId: taskId,
-            reminder: true,
-            automated: true,
-            read: false,
-            createdAt: admin.firestore.FieldValue.serverTimestamp(),
-            icon: 'fa-clock'
-          });
-          remindersCount++;
-        }
-      }
-
-      console.log(`✅ Sent ${overdueCount} overdue alerts and ${remindersCount} upcoming reminders`);
-      return { overdueCount, remindersCount };
-
-    } catch (error) {
-      console.error('❌ Error in dailyTaskReminders:', error);
-      throw error;
-    }
-  });
-
-/**
- * dailyBudgetWarnings - אזהרות תקציב יומיות
- * רץ כל יום בשעה 17:00 אחה"צ
- * בודק:
- * 1. משימות שחרגו מ-80% מתקציב הזמן (warning)
- * 2. משימות שחרגו 100% מתקציב הזמן (danger)
- * שולח התראה אוטומטית לעובדים
- */
-exports.dailyBudgetWarnings = onSchedule({
-  schedule: '0 17 * * *',  // כל יום בשעה 17:00
-  timeZone: 'Asia/Jerusalem',
-  region: 'us-central1'
-}, async (event) => {
-    try {
-      console.log('💰 Running dailyBudgetWarnings at', new Date().toISOString());
-
-      // מצא משימות פעילות
-      const tasksSnapshot = await db.collection('budget_tasks')
-        .where('status', '==', 'פעיל')
-        .get();
-
-      let warningsCount = 0;
-      let criticalCount = 0;
-
-      for (const taskDoc of tasksSnapshot.docs) {
-        const task = taskDoc.data();
-        const taskId = taskDoc.id;
-
-        // חישוב תקציב ושעות בפועל
-        const estimatedMinutes = (task.estimatedHours || 0) * 60 + (task.estimatedMinutes || 0);
-        const actualMinutes = (task.actualHours || 0) * 60 + (task.actualMinutes || 0);
-
-        // דלג על משימות ללא תקציב
-        if (estimatedMinutes === 0) continue;
-
-        const percentageUsed = (actualMinutes / estimatedMinutes) * 100;
-
-        // בדוק אם כבר שלחנו התראה היום (למנוע spam)
-        const today = new Date();
-        today.setHours(0, 0, 0, 0);
-        const existingNotification = await db.collection('notifications')
-          .where('taskId', '==', taskId)
-          .where('automated', '==', true)
-          .where('type', 'in', ['warning', 'error'])
-          .where('createdAt', '>=', admin.firestore.Timestamp.fromDate(today))
-          .limit(1)
-          .get();
-
-        if (!existingNotification.empty) {
-          console.log(`⏭️  Skipping task ${taskId} - already notified today`);
-          continue;
-        }
-
-        if (percentageUsed >= 100) {
-          // חריגה מלאה מהתקציב
-          await db.collection('notifications').add({
-            userId: task.lawyer || task.createdBy,
-            userEmail: task.employee,
-            title: `🚨 חריגה מתקציב: ${task.clientName}`,
-            message: `המשימה "${task.description}" חרגה מתקציב הזמן (${Math.round(percentageUsed)}%)`,
-            type: 'error',
-            taskId: taskId,
-            budgetWarning: true,
-            automated: true,
-            read: false,
-            createdAt: admin.firestore.FieldValue.serverTimestamp(),
-            icon: 'fa-exclamation-circle'
-          });
-          criticalCount++;
-
-        } else if (percentageUsed >= 80) {
-          // אזהרה - מתקרב לתקציב
-          await db.collection('notifications').add({
-            userId: task.lawyer || task.createdBy,
-            userEmail: task.employee,
-            title: `⚠️ התקרבות לתקציב: ${task.clientName}`,
-            message: `המשימה "${task.description}" מתקרבת לתקציב הזמן (${Math.round(percentageUsed)}%)`,
-            type: 'warning',
-            taskId: taskId,
-            budgetWarning: true,
-            automated: true,
-            read: false,
-            createdAt: admin.firestore.FieldValue.serverTimestamp(),
-            icon: 'fa-exclamation-triangle'
-          });
-          warningsCount++;
-        }
-      }
-
-      console.log(`✅ Sent ${criticalCount} critical budget alerts and ${warningsCount} budget warnings`);
-      return { criticalCount, warningsCount };
-
-    } catch (error) {
-      console.error('❌ Error in dailyBudgetWarnings:', error);
-      throw error;
-    }
-  });
-
-/**
- * formatDate - פורמט תאריך לתצוגה בעברית
- */
-function formatDate(date) {
-  return new Intl.DateTimeFormat('he-IL', {
-    year: 'numeric',
-    month: 'long',
-    day: 'numeric'
-  }).format(date);
-}
-
-// ===============================
-// Master Admin Panel Functions
-// ===============================
-
-// Import admin panel functions
-// ⚠️ TEMPORARILY DISABLED - admin directory not deployed
-// const { adminTransferUserData } = require('./admin/transfer-user-data');
-// const { adminGetUserFullDetails } = require('./admin/get-user-full-details');
-// const { adminGenerateClientReport } = require('./admin/generate-client-report');
-// const { adminUpdateClientFull } = require('./admin/update-client-full');
-
-// Import Master Admin Panel Phase 4 Wrappers (for Phase 3 UI)
-// ✅ ENABLED - Master Admin Panel Cloud Functions
-const {
-  createUser,
-  updateUser,
-  blockUser,
-  deleteUser,
-  getUserFullDetails
-} = require('./admin/master-admin-wrappers');
-
-// Export admin functions
-// ⚠️ TEMPORARILY DISABLED - admin directory not deployed
-// exports.adminTransferUserData = adminTransferUserData;
-// exports.adminGetUserFullDetails = adminGetUserFullDetails;
-// exports.adminGenerateClientReport = adminGenerateClientReport;
-// exports.adminUpdateClientFull = adminUpdateClientFull;
-
-// Export Master Admin Panel Phase 4 Wrappers (Simple names for UI)
-// ✅ ENABLED - Master Admin Panel Cloud Functions
-exports.createUser = createUser;
-exports.updateUser = updateUser;
-exports.blockUser = blockUser;
-exports.deleteUser = deleteUser;
-exports.getUserFullDetails = getUserFullDetails;
-
 // Real-time Task Updates & Notifications
 exports.updateBudgetTask = updateBudgetTask;
 exports.markNotificationAsRead = markNotificationAsRead;
-
-// ═══════════════════════════════════════════════════════════════════════
-// 📊 USER METRICS - Server-Side Statistics
-// ═══════════════════════════════════════════════════════════════════════
-
-/**
- * getUserMetrics - קבלת סטטיסטיקות משתמש מהשרת
- *
- * מחזיר מטריקות מחושבות מראש מ-user_metrics collection
- * אם לא קיים - מחשב בזמן אמת (fallback)
- *
- * @returns {Object} { total, active, completed, urgent, updatedAt }
- */
-exports.getUserMetrics = functions.https.onCall(async (data, context) => {
-  try {
-    // Security: בדיקת הרשאות
-    const user = await checkUserPermissions(context);
-
-    // נסה לקרוא metrics מראש מחושבים
-    const metricsDoc = await db.collection('user_metrics').doc(user.email).get();
-
-    if (metricsDoc.exists) {
-      const metrics = metricsDoc.data();
-
-      // בדוק שהנתונים לא ישנים מדי (יותר מ-5 דקות)
-      const now = Date.now();
-      const updatedAt = metrics.updatedAt?.toMillis() || 0;
-      const ageMinutes = (now - updatedAt) / (1000 * 60);
-
-      if (ageMinutes < 5) {
-        // נתונים טריים - החזר מיידית
-        return {
-          success: true,
-          data: {
-            total: metrics.total || 0,
-            active: metrics.active || 0,
-            completed: metrics.completed || 0,
-            urgent: metrics.urgent || 0,
-            updatedAt: metrics.updatedAt,
-            source: 'cache'
-          }
-        };
-      }
-    }
-
-    // Fallback: חישוב בזמן אמת
-    console.log(`⚡ Computing real-time metrics for ${user.email}`);
-
-    const tasksSnapshot = await db.collection('budget_tasks')
-      .where('employee', '==', user.email)
-      .get();
-
-    const now = new Date();
-    const urgentThresholdMs = 72 * 60 * 60 * 1000; // 72 hours
-
-    let total = 0;
-    let active = 0;
-    let completed = 0;
-    let urgent = 0;
-
-    tasksSnapshot.forEach(doc => {
-      const task = doc.data();
-      total++;
-
-      if (task.status === 'הושלם') {
-        completed++;
-      } else {
-        active++;
-
-        // בדוק דחיפות
-        if (task.deadline) {
-          const deadline = task.deadline.toMillis ? task.deadline.toMillis() : new Date(task.deadline).getTime();
-          const timeUntilDeadline = deadline - now.getTime();
-
-          if (timeUntilDeadline <= urgentThresholdMs && timeUntilDeadline >= -24 * 60 * 60 * 1000) {
-            urgent++;
-          }
-        }
-      }
-    });
-
-    const metrics = {
-      total,
-      active,
-      completed,
-      urgent,
-      updatedAt: admin.firestore.FieldValue.serverTimestamp()
-    };
-
-    // שמור לcache עבור פעם הבאה
-    await db.collection('user_metrics').doc(user.email).set(metrics, { merge: true });
-
-    return {
-      success: true,
-      data: {
-        ...metrics,
-        source: 'computed'
-      }
-    };
-
-  } catch (error) {
-    console.error('❌ getUserMetrics error:', error);
-    throw new functions.https.HttpsError(
-      'internal',
-      `שגיאה בטעינת מטריקות: ${error.message}`
-    );
-  }
-});
-
-/**
- * updateMetricsOnTaskChange - טריגר עדכון מטריקות על שינוי משימה
- *
- * מתעדכן אוטומטית כאשר:
- * - נוצרת משימה חדשה (onCreate)
- * - משימה משתנה (onUpdate)
- * - משימה נמחקת (onDelete)
- *
- * מעדכן את user_metrics/{email} באופן אטומי
- */
-exports.updateMetricsOnTaskChange = onDocumentWritten({
-  document: 'budget_tasks/{taskId}',
-  region: 'us-central1'
-}, async (event) => {
-    try {
-      const taskId = event.params.taskId;
-      const change = event.data;
-
-      // קבל את המשימה (לפני/אחרי)
-      const oldTask = change.before.exists ? change.before.data() : null;
-      const newTask = change.after.exists ? change.after.data() : null;
-
-      // אם אין employee - דלג
-      const employee = (newTask?.employee || oldTask?.employee);
-      if (!employee) {
-        console.log(`⏭️  Skipping task ${taskId} - no employee`);
-        return null;
-      }
-
-      console.log(`📊 Updating metrics for ${employee} (task: ${taskId})`);
-
-      // חשב שינוי במטריקות
-      const metricsRef = db.collection('user_metrics').doc(employee);
-
-      const now = new Date();
-      const urgentThresholdMs = 72 * 60 * 60 * 1000; // 72 hours
-
-      // פונקציה לבדיקת דחיפות
-      const isUrgent = (task) => {
-        if (!task?.deadline || task.status === 'הושלם') return false;
-        const deadline = task.deadline.toMillis ? task.deadline.toMillis() : new Date(task.deadline).getTime();
-        const timeUntilDeadline = deadline - now.getTime();
-        return timeUntilDeadline <= urgentThresholdMs && timeUntilDeadline >= -24 * 60 * 60 * 1000;
-      };
-
-      // חשב שינויים
-      let totalDelta = 0;
-      let activeDelta = 0;
-      let completedDelta = 0;
-      let urgentDelta = 0;
-
-      if (!oldTask && newTask) {
-        // משימה חדשה
-        totalDelta = 1;
-        if (newTask.status === 'הושלם') {
-          completedDelta = 1;
-        } else {
-          activeDelta = 1;
-          if (isUrgent(newTask)) urgentDelta = 1;
-        }
-      } else if (oldTask && !newTask) {
-        // משימה נמחקה
-        totalDelta = -1;
-        if (oldTask.status === 'הושלם') {
-          completedDelta = -1;
-        } else {
-          activeDelta = -1;
-          if (isUrgent(oldTask)) urgentDelta = -1;
-        }
-      } else if (oldTask && newTask) {
-        // משימה השתנתה
-        const oldCompleted = oldTask.status === 'הושלם';
-        const newCompleted = newTask.status === 'הושלם';
-        const oldUrgent = isUrgent(oldTask);
-        const newUrgent = isUrgent(newTask);
-
-        if (oldCompleted !== newCompleted) {
-          if (newCompleted) {
-            activeDelta = -1;
-            completedDelta = 1;
-            if (oldUrgent) urgentDelta = -1;
-          } else {
-            activeDelta = 1;
-            completedDelta = -1;
-            if (newUrgent) urgentDelta = 1;
-          }
-        } else if (!newCompleted && oldUrgent !== newUrgent) {
-          // שינוי בדחיפות (בלי שינוי סטטוס)
-          urgentDelta = newUrgent ? 1 : -1;
-        }
-      }
-
-      // עדכון אטומי
-      if (totalDelta !== 0 || activeDelta !== 0 || completedDelta !== 0 || urgentDelta !== 0) {
-        await metricsRef.set({
-          total: admin.firestore.FieldValue.increment(totalDelta),
-          active: admin.firestore.FieldValue.increment(activeDelta),
-          completed: admin.firestore.FieldValue.increment(completedDelta),
-          urgent: admin.firestore.FieldValue.increment(urgentDelta),
-          updatedAt: admin.firestore.FieldValue.serverTimestamp()
-        }, { merge: true });
-
-        console.log(`✅ Metrics updated: total(${totalDelta > 0 ? '+' : ''}${totalDelta}) active(${activeDelta > 0 ? '+' : ''}${activeDelta}) completed(${completedDelta > 0 ? '+' : ''}${completedDelta}) urgent(${urgentDelta > 0 ? '+' : ''}${urgentDelta})`);
-      } else {
-        console.log(`⏭️  No metric changes for task ${taskId}`);
-      }
-
-      return null;
-    } catch (error) {
-      console.error('❌ updateMetricsOnTaskChange error:', error);
-      // לא נזרוק שגיאה - טריגר לא צריך לעצור פעולות
-      return null;
-    }
-  });
-
-// ═══════════════════════════════════════════════════════════════════════
-// 🚨 NUCLEAR CLEANUP - Admin Only
-// ═══════════════════════════════════════════════════════════════════════
-// ⚠️ TEMPORARILY DISABLED - admin directory not deployed
-// const { nuclearCleanup } = require('./admin/nuclear-cleanup');
-// exports.nuclearCleanup = nuclearCleanup;
 
 // ═══════════════════════════════════════════════════════════════════════
 // 🔐 SET ADMIN CLAIMS - One-Time Setup Function
@@ -6160,1379 +5523,7 @@ exports.setAdminClaims = functions.https.onRequest(async (req, res) => {
   });
 });
 
-// ===============================
-// Task Approval System
-// ===============================
 
-/**
- * Approve task budget request
- */
-exports.approveTaskBudget = functions.https.onCall(async (data, context) => {
-  try {
-    const user = await checkUserPermissions(context);
-
-    // Only admins can approve
-    const tokenResult = await auth.getUser(context.auth.uid);
-    const isAdmin = tokenResult.customClaims?.role === 'admin';
-
-    if (!isAdmin) {
-      throw new functions.https.HttpsError(
-        'permission-denied',
-        'רק מנהלים יכולים לאשר תקציבים'
-      );
-    }
-
-    const { approvalId, approvedMinutes, adminNotes } = data;
-
-    if (!approvalId) {
-      throw new functions.https.HttpsError('invalid-argument', 'חסר מזהה בקשה');
-    }
-
-    if (!approvedMinutes || approvedMinutes <= 0) {
-      throw new functions.https.HttpsError('invalid-argument', 'תקציב מאושר חייב להיות חיובי');
-    }
-
-    // Get approval request
-    const approvalRef = db.collection('pending_task_approvals').doc(approvalId);
-    const approvalDoc = await approvalRef.get();
-
-    if (!approvalDoc.exists) {
-      throw new functions.https.HttpsError('not-found', 'בקשת אישור לא נמצאה');
-    }
-
-    const approval = approvalDoc.data();
-    const taskId = approval.taskId;
-    const requestedMinutes = approval.requestedMinutes || approval.taskData.estimatedMinutes || 0;
-    const isModified = approvedMinutes !== requestedMinutes;
-    const newStatus = isModified ? 'modified' : 'approved';
-
-    // Use batch for atomic update
-    const batch = db.batch();
-
-    // Update approval status
-    batch.update(approvalRef, {
-      status: newStatus,
-      reviewedBy: user.email,
-      reviewedByName: user.username,
-      reviewedAt: admin.firestore.FieldValue.serverTimestamp(),
-      approvedMinutes: approvedMinutes,
-      adminNotes: adminNotes || ''
-    });
-
-    // Update task
-    const taskRef = db.collection('budget_tasks').doc(taskId);
-    batch.update(taskRef, {
-      status: 'פעיל',
-      estimatedMinutes: approvedMinutes,
-      estimatedHours: approvedMinutes / 60,
-      approvedMinutes: approvedMinutes,
-      approvedBy: user.email,
-      approvedAt: admin.firestore.FieldValue.serverTimestamp()
-    });
-
-    // Create notification message
-    const messageText = isModified
-      ? `✅ תקציב המשימה אושר עם שינוי\n\n📋 משימה: ${approval.taskData.description}\n⏱️ תקציב מבוקש: ${requestedMinutes} דקות\n✅ תקציב מאושר: ${approvedMinutes} דקות${adminNotes ? `\n📝 הערות: ${adminNotes}` : ''}`
-      : `✅ תקציב המשימה אושר במלואו\n\n📋 משימה: ${approval.taskData.description}\n⏱️ תקציב: ${approvedMinutes} דקות`;
-
-    const messageRef = db.collection('user_messages').doc();
-    batch.set(messageRef, {
-      to: approval.requestedBy,
-      from: 'system',
-      fromName: 'מערכת',
-      message: messageText,
-      type: 'task_approval',
-      taskId: taskId,
-      approvalId: approvalId,
-      status: 'unread',
-      createdAt: admin.firestore.FieldValue.serverTimestamp()
-    });
-
-    await batch.commit();
-
-    console.log(`✅ Task ${taskId} approved: ${approvedMinutes} minutes`);
-
-    return {
-      success: true,
-      taskId: taskId,
-      status: newStatus
-    };
-
-  } catch (error) {
-    console.error('❌ Error approving task:', error);
-    throw error;
-  }
-});
-
-/**
- * ✅ OPTIMIZATION: Firestore Trigger for WhatsApp notifications
- * Automatically sends WhatsApp when a new approval request is created
- * This removes the 8-second blocking call from the frontend
- */
-exports.onApprovalCreated = onDocumentWritten(
-  'pending_task_approvals/{approvalId}',
-  async (event) => {
-    try {
-      const newData = event.data.after.data();
-      const oldData = event.data.before.data();
-
-      // Only trigger on new documents (create), not updates
-      if (oldData) {
-        console.log('⏭️ Skipping - document updated, not created');
-        return null;
-      }
-
-      // Only send WhatsApp for pending approvals
-      if (!newData || newData.status !== 'pending') {
-        console.log('⏭️ Skipping - status is not pending');
-        return null;
-      }
-
-      console.log(`📱 Sending WhatsApp for approval ${event.params.approvalId}`);
-
-      // Get all admins with WhatsApp enabled
-      const adminsSnapshot = await db.collection('employees')
-        .where('role', '==', 'admin')
-        .where('whatsappEnabled', '==', true)
-        .get();
-
-      if (adminsSnapshot.empty) {
-        console.log('⚠️ No admins with WhatsApp enabled');
-        return null;
-      }
-
-      // Initialize Twilio
-      // ✅ Use environment variables (v2 compatible) instead of functions.config()
-      const accountSid = process.env.TWILIO_ACCOUNT_SID || TWILIO_ACCOUNT_SID;
-      const authToken = process.env.TWILIO_AUTH_TOKEN || TWILIO_AUTH_TOKEN;
-      const fromNumber = process.env.TWILIO_WHATSAPP_NUMBER || TWILIO_WHATSAPP_NUMBER;
-
-      if (!accountSid || !authToken) {
-        console.error('❌ Twilio not configured');
-        return null;
-      }
-
-      const twilio = require('twilio');
-      const client = twilio(accountSid, authToken);
-
-      let sentCount = 0;
-
-      // Send to each admin
-      for (const adminDoc of adminsSnapshot.docs) {
-        const admin = adminDoc.data();
-
-        // Format phone number
-        let phone = (admin.phone || '').replace(/\D/g, '');
-        if (phone.startsWith('05')) {
-          phone = '972' + phone.substring(1);
-        } else if (!phone.startsWith('972')) {
-          phone = '972' + phone;
-        }
-        const toNumber = `whatsapp:+${phone}`;
-
-        // Calculate time display
-        const minutes = parseInt(newData.requestedMinutes) || parseInt(newData.taskData?.estimatedMinutes) || 0;
-        const hours = Math.floor(minutes / 60);
-        const mins = minutes % 60;
-        const timeStr = hours > 0
-          ? `${hours} שעות${mins > 0 ? ` ו-${mins} דקות` : ''}`
-          : `${mins} דקות`;
-
-        // Create message
-        const message = `🔔 משימה חדשה נוספה
-
-👤 ${newData.requestedByName || newData.requestedBy} הוסיף משימה:
-
-📋 לקוח: ${newData.taskData?.clientName || 'לא צוין'}
-📝 תיאור: ${newData.taskData?.description || 'לא צוין'}
-⏱️ תקציב: ${timeStr} (${minutes} דקות)
-
-━━━━━━━━━━━━━━━━━━━━
-
-ℹ️ לידיעה בלבד
-המשימה כבר פעילה והמשתמש יכול להתחיל לעבוד
-
-🤖 הודעה אוטומטית ממערכת ניהול`;
-
-        try {
-          await client.messages.create({
-            from: fromNumber,
-            to: toNumber,
-            body: message
-          });
-          sentCount++;
-          console.log(`✅ WhatsApp sent to ${admin.username || admin.name}`);
-        } catch (smsError) {
-          console.error(`❌ Failed to send WhatsApp to ${admin.username}:`, smsError.message);
-        }
-      }
-
-      console.log(`✅ Trigger completed: ${sentCount} WhatsApp messages sent`);
-      return { success: true, sent: sentCount };
-
-    } catch (error) {
-      console.error('❌ Error in onApprovalCreated trigger:', error);
-      // Don't throw - we don't want to fail the approval creation
-      return null;
-    }
-  }
-);
-
-/**
- * Reject task budget request
- */
-exports.rejectTaskBudget = functions.https.onCall(async (data, context) => {
-  try {
-    const user = await checkUserPermissions(context);
-
-    // Only admins can reject
-    const tokenResult = await auth.getUser(context.auth.uid);
-    const isAdmin = tokenResult.customClaims?.role === 'admin';
-
-    if (!isAdmin) {
-      throw new functions.https.HttpsError(
-        'permission-denied',
-        'רק מנהלים יכולים לדחות תקציבים'
-      );
-    }
-
-    const { approvalId, rejectionReason } = data;
-
-    if (!approvalId) {
-      throw new functions.https.HttpsError('invalid-argument', 'חסר מזהה בקשה');
-    }
-
-    if (!rejectionReason || rejectionReason.trim().length < 3) {
-      throw new functions.https.HttpsError('invalid-argument', 'חובה להזין סיבת דחייה');
-    }
-
-    // Get approval request
-    const approvalRef = db.collection('pending_task_approvals').doc(approvalId);
-    const approvalDoc = await approvalRef.get();
-
-    if (!approvalDoc.exists) {
-      throw new functions.https.HttpsError('not-found', 'בקשת אישור לא נמצאה');
-    }
-
-    const approval = approvalDoc.data();
-    const taskId = approval.taskId;
-    const requestedMinutes = approval.requestedMinutes || approval.taskData.estimatedMinutes || 0;
-
-    // Use batch for atomic update
-    const batch = db.batch();
-
-    // Update approval status
-    batch.update(approvalRef, {
-      status: 'rejected',
-      reviewedBy: user.email,
-      reviewedByName: user.username,
-      reviewedAt: admin.firestore.FieldValue.serverTimestamp(),
-      rejectionReason: rejectionReason
-    });
-
-    // Delete the task (it was never approved)
-    const taskRef = db.collection('budget_tasks').doc(taskId);
-    batch.delete(taskRef);
-
-    // Create notification message
-    const messageText = `❌ בקשת תקציב נדחתה\n\n📋 משימה: ${approval.taskData.description}\n⏱️ תקציב מבוקש: ${requestedMinutes} דקות\n📝 סיבת דחייה: ${rejectionReason}`;
-
-    const messageRef = db.collection('user_messages').doc();
-    batch.set(messageRef, {
-      to: approval.requestedBy,
-      from: 'system',
-      fromName: 'מערכת',
-      message: messageText,
-      type: 'task_rejection',
-      taskId: taskId,
-      approvalId: approvalId,
-      status: 'unread',
-      createdAt: admin.firestore.FieldValue.serverTimestamp()
-    });
-
-    await batch.commit();
-
-    console.log(`✅ Task ${taskId} rejected`);
-
-    return {
-      success: true,
-      taskId: taskId
-    };
-
-  } catch (error) {
-    console.error('❌ Error rejecting task:', error);
-    throw error;
-  }
-});
-
-// ═══════════════════════════════════════════════════════════════════════════
-// DEPRECATED FUNCTIONS - Tasks are now auto-approved
-// ═══════════════════════════════════════════════════════════════════════════
-
-/**
- * @deprecated Tasks are now auto-approved - no manual approval needed
- * This function is kept for backward compatibility only
- * Tasks created with status: 'פעיל' automatically
- */
-exports.approveTaskBudget_DEPRECATED = exports.approveTaskBudget;
-exports.approveTaskBudget = functions.https.onCall(async (data, context) => {
-  console.warn('⚠️ approveTaskBudget is deprecated - tasks are auto-approved');
-  throw new functions.https.HttpsError(
-    'unimplemented',
-    'פונקציה זו אינה בשימוש יותר - משימות מאושרות אוטומטית'
-  );
-});
-
-/**
- * @deprecated Tasks are now auto-approved - no manual rejection needed
- * This function is kept for backward compatibility only
- */
-exports.rejectTaskBudget_DEPRECATED = exports.rejectTaskBudget;
-exports.rejectTaskBudget = functions.https.onCall(async (data, context) => {
-  console.warn('⚠️ rejectTaskBudget is deprecated - tasks are auto-approved');
-  throw new functions.https.HttpsError(
-    'unimplemented',
-    'פונקציה זו אינה בשימוש יותר - משימות מאושרות אוטומטית'
-  );
-});
-
-// ===============================
-// WhatsApp Broadcast with Twilio
-// ===============================
-
-/**
- * Send WhatsApp broadcast messages to selected employees
- * Uses Twilio WhatsApp Business API
- */
-exports.sendBroadcastMessage = functions.https.onCall(async (data, context) => {
-  try {
-    // Check authentication
-    if (!context.auth) {
-      throw new functions.https.HttpsError(
-        'unauthenticated',
-        'נדרשת התחברות למערכת'
-      );
-    }
-
-    // Check if user is admin
-    const userEmail = context.auth.token.email;
-    const employeeDoc = await db.collection('employees').doc(userEmail).get();
-
-    if (!employeeDoc.exists || employeeDoc.data().role !== 'admin') {
-      throw new functions.https.HttpsError(
-        'permission-denied',
-        'רק מנהלים יכולים לשלוח הודעות broadcast'
-      );
-    }
-
-    // Validate input
-    const { employeeEmails, templateType, customMessage } = data;
-
-    if (!employeeEmails || !Array.isArray(employeeEmails) || employeeEmails.length === 0) {
-      throw new functions.https.HttpsError(
-        'invalid-argument',
-        'חייב לספק רשימת עובדים'
-      );
-    }
-
-    if (!templateType) {
-      throw new functions.https.HttpsError(
-        'invalid-argument',
-        'חייב לבחור תבנית הודעה'
-      );
-    }
-
-    // Initialize Twilio (get credentials from Firebase Config)
-    const twilioConfig = functions.config().twilio;
-
-    if (!twilioConfig || !twilioConfig.account_sid || !twilioConfig.auth_token) {
-      throw new functions.https.HttpsError(
-        'failed-precondition',
-        'Twilio לא מוגדר. הרץ: firebase functions:config:set twilio.account_sid="YOUR_SID" twilio.auth_token="YOUR_TOKEN" twilio.whatsapp_number="whatsapp:+14155238886"'
-      );
-    }
-
-    const twilio = require('twilio');
-    const client = twilio(twilioConfig.account_sid, twilioConfig.auth_token);
-    const fromNumber = twilioConfig.whatsapp_number || 'whatsapp:+14155238886'; // Twilio Sandbox default
-
-    // Message templates
-    const templates = {
-      DAILY_REMINDER: (name) => `שלום ${name}! ⏰\n\nתזכורת לרישום שעות היום במערכת.\n\nכניסה למערכת:\nhttps://gh-law-office-system.netlify.app`,
-
-      WEEKLY_SUMMARY: (name) => `שלום ${name}! 📅\n\nבקשה לעדכן את סיכום שעות השבוע במערכת.\n\nכניסה למערכת:\nhttps://gh-law-office-system.netlify.app`,
-
-      SYSTEM_ANNOUNCEMENT: (name, message) => `שלום ${name}! 📢\n\nהודעת מערכת:\n${message}\n\nכניסה למערכת:\nhttps://gh-law-office-system.netlify.app`,
-
-      CUSTOM: (name, message) => `שלום ${name}!\n\n${message}\n\nכניסה למערכת:\nhttps://gh-law-office-system.netlify.app`
-    };
-
-    // Results tracking
-    const results = {
-      success: [],
-      failed: []
-    };
-
-    // Send messages to each employee
-    for (const email of employeeEmails) {
-      try {
-        // Get employee data
-        const empDoc = await db.collection('employees').doc(email).get();
-
-        if (!empDoc.exists) {
-          results.failed.push({
-            email,
-            name: email,
-            error: 'עובד לא נמצא במערכת'
-          });
-          continue;
-        }
-
-        const employee = empDoc.data();
-        const name = employee.name || employee.username || email;
-
-        // Check if employee has WhatsApp enabled and phone number
-        if (!employee.whatsappEnabled || !employee.phone) {
-          results.failed.push({
-            email,
-            name,
-            error: 'WhatsApp לא מופעל או אין מספר טלפון'
-          });
-          continue;
-        }
-
-        // Format phone number for WhatsApp
-        let phone = employee.phone.replace(/\D/g, ''); // Remove non-digits
-
-        // Israeli phone format: 05X-XXXXXXX -> +9725XXXXXXXX
-        if (phone.startsWith('05')) {
-          phone = '972' + phone.substring(1);
-        } else if (!phone.startsWith('972')) {
-          phone = '972' + phone;
-        }
-
-        const toNumber = `whatsapp:+${phone}`;
-
-        // Generate message
-        let messageBody;
-        if (templateType === 'SYSTEM_ANNOUNCEMENT' || templateType === 'CUSTOM') {
-          messageBody = templates[templateType](name, customMessage);
-        } else {
-          messageBody = templates[templateType](name);
-        }
-
-        // Send via Twilio
-        const message = await client.messages.create({
-          from: fromNumber,
-          to: toNumber,
-          body: messageBody
-        });
-
-        results.success.push({
-          email,
-          name,
-          phone: toNumber,
-          messageSid: message.sid
-        });
-
-        console.log(`✅ WhatsApp sent to ${name} (${email}): ${message.sid}`);
-
-      } catch (error) {
-        console.error(`❌ Failed to send to ${email}:`, error);
-        results.failed.push({
-          email,
-          name: email,
-          error: error.message || 'שגיאה בשליחה'
-        });
-      }
-    }
-
-    // Log to audit
-    await logAction('whatsapp_broadcast', context.auth.uid, userEmail, {
-      templateType,
-      totalSent: results.success.length,
-      totalFailed: results.failed.length,
-      recipients: employeeEmails
-    });
-
-    return {
-      totalSent: results.success.length,
-      totalFailed: results.failed.length,
-      results
-    };
-
-  } catch (error) {
-    console.error('❌ sendBroadcastMessage error:', error);
-    throw new functions.https.HttpsError(
-      'internal',
-      error.message || 'שגיאה בשליחת הודעות'
-    );
-  }
-});
-
-// ===============================
-// WhatsApp Task Approval Automation
-// ===============================
-
-/**
- * Send WhatsApp notification to admin when new task approval is requested
- * Called automatically when approval is created
- */
-exports.sendWhatsAppApprovalNotification = functions.https.onCall(async (data, context) => {
-  try {
-    // Check authentication
-    if (!context.auth) {
-      throw new functions.https.HttpsError('unauthenticated', 'נדרשת התחברות');
-    }
-
-    const { approvalId, taskData, requestedBy, requestedByName } = data;
-
-    if (!approvalId || !taskData) {
-      throw new functions.https.HttpsError('invalid-argument', 'חסרים פרמטרים');
-    }
-
-    // Get all admins with WhatsApp enabled
-    const adminsSnapshot = await db.collection('employees')
-      .where('role', '==', 'admin')
-      .where('whatsappEnabled', '==', true)
-      .get();
-
-    if (adminsSnapshot.empty) {
-      console.log('⚠️ No admins with WhatsApp enabled');
-      return { success: true, sent: 0, message: 'אין מנהלים עם WhatsApp מופעל' };
-    }
-
-    // Initialize Twilio
-    const twilioConfig = functions.config().twilio;
-    if (!twilioConfig?.account_sid || !twilioConfig?.auth_token) {
-      throw new functions.https.HttpsError('failed-precondition', 'Twilio לא מוגדר');
-    }
-
-    const twilio = require('twilio');
-    const client = twilio(twilioConfig.account_sid, twilioConfig.auth_token);
-    const fromNumber = twilioConfig.whatsapp_number || 'whatsapp:+14155238886';
-
-    const results = [];
-
-    // Send to each admin
-    for (const adminDoc of adminsSnapshot.docs) {
-      const admin = adminDoc.data();
-      const adminName = admin.name || admin.username || adminDoc.id;
-
-      // Format phone number
-      let phone = (admin.phone || '').replace(/\D/g, '');
-      if (phone.startsWith('05')) {
-        phone = '972' + phone.substring(1);
-      } else if (!phone.startsWith('972')) {
-        phone = '972' + phone;
-      }
-      const toNumber = `whatsapp:+${phone}`;
-
-      // Calculate hours
-      const minutes = parseInt(taskData.estimatedMinutes) || 0;
-      const hours = Math.floor(minutes / 60);
-      const mins = minutes % 60;
-      const timeStr = hours > 0
-        ? `${hours} שעות${mins > 0 ? ` ו-${mins} דקות` : ''}`
-        : `${mins} דקות`;
-
-      // Create message
-      const message = `🔔 משימה חדשה לאישור
-
-👤 ${requestedByName || requestedBy} מבקש אישור תקציב:
-
-📋 לקוח: ${taskData.clientName || 'לא צוין'}
-📝 תיאור: ${taskData.description}
-⏱️ תקציב: ${timeStr} (${minutes} דקות)
-
-━━━━━━━━━━━━━━━━━━━━
-
-📲 לאישור - כתוב:
-✅ "אישור" - לאשר כמו שביקש
-✅ "אישור 90" - לאשר עם 90 דקות
-
-📲 לדחייה - כתוב:
-❌ "דחייה" + סיבה
-דוגמה: "דחייה תקציב גבוה"
-
-💡 כתוב "משימות" לראות הכל
-
-🤖 הודעה אוטומטית ממערכת ניהול`;
-
-      try {
-        const twilioMessage = await client.messages.create({
-          from: fromNumber,
-          to: toNumber,
-          body: message
-        });
-
-        results.push({
-          admin: adminName,
-          phone: toNumber,
-          success: true,
-          messageSid: twilioMessage.sid
-        });
-
-        console.log(`✅ Approval notification sent to ${adminName}: ${twilioMessage.sid}`);
-
-      } catch (error) {
-        console.error(`❌ Failed to send to ${adminName}:`, error);
-        results.push({
-          admin: adminName,
-          phone: toNumber,
-          success: false,
-          error: error.message
-        });
-      }
-    }
-
-    // Save notification log
-    await db.collection('whatsapp_approval_notifications').add({
-      approvalId,
-      taskId: taskData.taskId || null,
-      requestedBy,
-      sentTo: results.map(r => r.admin),
-      sentAt: admin.firestore.FieldValue.serverTimestamp(),
-      results
-    });
-
-    const successCount = results.filter(r => r.success).length;
-
-    return {
-      success: true,
-      sent: successCount,
-      total: results.length,
-      results
-    };
-
-  } catch (error) {
-    console.error('❌ sendWhatsAppApprovalNotification error:', error);
-    throw new functions.https.HttpsError('internal', error.message || 'שגיאה בשליחת התראה');
-  }
-});
-
-/**
- * Webhook to receive WhatsApp messages from Twilio
- * Handles approval/rejection responses from admins
- */
-exports.whatsappWebhook = onRequest({
-  region: 'us-central1'
-}, async (req, res) => {
-  try {
-    // Get message data
-    const { From, Body, MessageSid, NumMedia, MediaUrl0, MediaContentType0 } = req.body;
-
-    console.log(`📨 WhatsApp message received from ${From}: "${Body}"`);
-    console.log(`📎 Media: NumMedia=${NumMedia}, Type=${MediaContentType0}`);
-
-    if (!From) {
-      res.status(400).send('Missing From parameter');
-      return;
-    }
-
-    // Extract phone number
-    const phoneNumber = From.replace('whatsapp:', '').replace('+', '');
-
-    // Import the WhatsApp Bot
-    const WhatsAppBot = require('./src/whatsapp-bot/WhatsAppBot');
-    const bot = new WhatsAppBot();
-
-    // Identify user
-    const userInfo = await bot.identifyUser(phoneNumber);
-
-    // Only allow admins to use the bot
-    if (userInfo.role !== 'admin') {
-      console.log(`⚠️ Message from non-admin: ${From} (${userInfo.name})`);
-      res.status(200).send('OK');
-      return;
-    }
-
-    console.log(`✅ Admin identified: ${userInfo.name || userInfo.email}`);
-
-    // Check if this is a media message
-    const hasMedia = NumMedia && parseInt(NumMedia) > 0;
-    let response;
-
-    if (hasMedia && MediaUrl0) {
-      // Handle media message (PDF upload)
-      console.log(`📎 Processing media message: ${MediaContentType0}`);
-      response = await bot.handleMediaMessage(
-        phoneNumber,
-        MediaUrl0,
-        MediaContentType0,
-        Body || '',
-        userInfo
-      );
-    } else if (Body) {
-      // Handle regular text message
-      response = await bot.handleMessage(phoneNumber, Body, userInfo);
-    } else {
-      res.status(400).send('Missing Body or Media');
-      return;
-    }
-
-    // Send response via Twilio
-    const twilioAccountSid = process.env.TWILIO_ACCOUNT_SID || 'AC9e5e9e3c953a5bbb878622b6e70201b6';
-    const twilioAuthToken = process.env.TWILIO_AUTH_TOKEN || 'fed2170530e4ed34d3b1b3407e0f0f5f';
-    const twilioWhatsappNumber = process.env.TWILIO_WHATSAPP_NUMBER || 'whatsapp:+14155238886';
-
-    if (twilioAccountSid && response) {
-      const twilio = require('twilio');
-      const client = twilio(twilioAccountSid, twilioAuthToken);
-
-      // ═══ בדוק אם צריך לשלוח Template או טקסט רגיל ═══
-      if (typeof response === 'object' && response.useTemplate) {
-        // שלח Content Template עם כפתורים
-        console.log(`📤 Sending template: ${response.templateSid}`);
-        await client.messages.create({
-          from: twilioWhatsappNumber,
-          to: From,
-          contentSid: response.templateSid,
-          contentVariables: JSON.stringify(response.variables)
-        });
-        console.log(`✅ Template sent to ${userInfo.name}`);
-      } else {
-        // שלח הודעת טקסט רגילה
-        await client.messages.create({
-          from: twilioWhatsappNumber,
-          to: From,
-          body: typeof response === 'string' ? response : JSON.stringify(response)
-        });
-        console.log(`✅ Bot response sent to ${userInfo.name}`);
-      }
-    }
-
-    // Log the interaction
-    await db.collection('whatsapp_bot_interactions').add({
-      from: From,
-      userId: userInfo.email,
-      userName: userInfo.name || userInfo.email,
-      message: Body,
-      response: typeof response === 'object' && response.useTemplate
-        ? `[Template: ${response.templateSid}]`
-        : response,
-      messageSid: MessageSid,
-      timestamp: admin.firestore.FieldValue.serverTimestamp()
-    });
-
-    res.status(200).send('OK');
-
-  } catch (error) {
-    console.error('❌ whatsappWebhook error:', error);
-
-    // Try to send error message to user
-    try {
-      const { From } = req.body;
-      const twilioAccountSid = process.env.TWILIO_ACCOUNT_SID || 'AC9e5e9e3c953a5bbb878622b6e70201b6';
-      const twilioAuthToken = process.env.TWILIO_AUTH_TOKEN || 'fed2170530e4ed34d3b1b3407e0f0f5f';
-      const twilioWhatsappNumber = process.env.TWILIO_WHATSAPP_NUMBER || 'whatsapp:+14155238886';
-
-      if (twilioAccountSid && From) {
-        const twilio = require('twilio');
-        const client = twilio(twilioAccountSid, twilioAuthToken);
-        await client.messages.create({
-          from: twilioWhatsappNumber,
-          to: From,
-          body: '❌ מצטער, הייתה שגיאה במערכת. נסה שוב מאוחר יותר או כתוב "עזרה"'
-        });
-      }
-    } catch (sendError) {
-      console.error('❌ Failed to send error message:', sendError);
-    }
-
-    res.status(500).send('Error');
-  }
-});
-
-// ===============================
-// DELETE USER DATA - מחיקת נתוני משתמש
-// ===============================
-
-/**
- * Delete user data (tasks, timesheets, approvals)
- * מחיקת נתוני משתמש (משימות, שעתונים, אישורים)
- */
-exports.deleteUserData = functions.https.onCall(async (data, context) => {
-  try {
-    // Check authentication
-    if (!context.auth) {
-      throw new functions.https.HttpsError('unauthenticated', 'נדרשת התחברות');
-    }
-
-    // Check admin permissions
-    const callerEmail = context.auth.token.email;
-    const adminSnapshot = await db.collection('employees').where('email', '==', callerEmail).get();
-
-    if (adminSnapshot.empty) {
-      throw new functions.https.HttpsError('permission-denied', 'אין הרשאות מנהל');
-    }
-
-    const adminData = adminSnapshot.docs[0].data();
-    if (!adminData.isAdmin && adminData.role !== 'admin') {
-      throw new functions.https.HttpsError('permission-denied', 'אין הרשאות מנהל');
-    }
-
-    const { email, deleteTasks, deleteTimesheets, deleteApprovals } = data;
-
-    if (!email) {
-      throw new functions.https.HttpsError('invalid-argument', 'חסר אימייל משתמש');
-    }
-
-    console.log(`🗑️ Deleting user data for: ${email}`);
-    console.log(`   Tasks: ${deleteTasks}, Timesheets: ${deleteTimesheets}, Approvals: ${deleteApprovals}`);
-
-    let deletedCounts = {
-      tasks: 0,
-      timesheets: 0,
-      approvals: 0
-    };
-
-    // Delete budget_tasks
-    if (deleteTasks) {
-      const tasksQuery = db.collection('budget_tasks').where('employeeEmail', '==', email);
-      let tasksSnapshot = await tasksQuery.get();
-
-      while (!tasksSnapshot.empty) {
-        const batch = db.batch();
-        tasksSnapshot.docs.forEach(doc => {
-          batch.delete(doc.ref);
-          deletedCounts.tasks++;
-        });
-        await batch.commit();
-
-        // Get next batch
-        tasksSnapshot = await tasksQuery.limit(500).get();
-      }
-      console.log(`✅ Deleted ${deletedCounts.tasks} tasks`);
-    }
-
-    // Delete timesheet_entries
-    if (deleteTimesheets) {
-      const timesheetsQuery = db.collection('timesheet_entries').where('employeeEmail', '==', email);
-      let timesheetsSnapshot = await timesheetsQuery.get();
-
-      while (!timesheetsSnapshot.empty) {
-        const batch = db.batch();
-        timesheetsSnapshot.docs.forEach(doc => {
-          batch.delete(doc.ref);
-          deletedCounts.timesheets++;
-        });
-        await batch.commit();
-
-        // Get next batch
-        timesheetsSnapshot = await timesheetsQuery.limit(500).get();
-      }
-      console.log(`✅ Deleted ${deletedCounts.timesheets} timesheet entries`);
-    }
-
-    // Delete pending_task_approvals
-    if (deleteApprovals) {
-      const approvalsQuery = db.collection('pending_task_approvals').where('requestedBy', '==', email);
-      let approvalsSnapshot = await approvalsQuery.get();
-
-      while (!approvalsSnapshot.empty) {
-        const batch = db.batch();
-        approvalsSnapshot.docs.forEach(doc => {
-          batch.delete(doc.ref);
-          deletedCounts.approvals++;
-        });
-        await batch.commit();
-
-        // Get next batch
-        approvalsSnapshot = await approvalsQuery.limit(500).get();
-      }
-      console.log(`✅ Deleted ${deletedCounts.approvals} task approvals`);
-    }
-
-    // Log the action
-    await db.collection('audit_log').add({
-      timestamp: admin.firestore.FieldValue.serverTimestamp(),
-      action: 'delete_user_data',
-      adminEmail: callerEmail,
-      targetEmail: email,
-      deletedCounts,
-      details: { deleteTasks, deleteTimesheets, deleteApprovals }
-    });
-
-    console.log(`✅ User data deleted successfully for: ${email}`);
-
-    return {
-      success: true,
-      message: 'הנתונים נמחקו בהצלחה',
-      deletedCounts
-    };
-
-  } catch (error) {
-    console.error('❌ Error deleting user data:', error);
-    throw new functions.https.HttpsError(
-      'internal',
-      error.message || 'שגיאה במחיקת נתונים'
-    );
-  }
-});
-
-// ===============================
-// DELETE USER DATA SELECTIVE - מחיקה סלקטיבית מאובטחת
-// 🔒 PHASE 1: READ-ONLY MODE
-// ===============================
-
-const { validateDeletionRequest, checkRateLimit } = require('./src/deletion/validators');
-const { verifyAllOwnership } = require('./src/deletion/ownership');
-const { executeDeletion, DELETION_ENABLED } = require('./src/deletion/deletion-engine');
-const { logDeletionAttempt, checkSuspiciousActivity } = require('./src/deletion/audit');
-
-/**
- * Delete user data selectively (tasks, timesheets, approvals)
- * מחיקה סלקטיבית של נתוני משתמש
- *
- * 🔒 Security Layers:
- * 1. Authentication - אימות
- * 2. Authorization - הרשאות admin בלבד
- * 3. Input Validation - וולידציה מלאה
- * 4. Rate Limiting - מניעת שימוש לרעה
- * 5. Ownership Verification - בדיקת בעלות
- * 6. Transaction Safety - מחיקה מאובטחת
- * 7. Audit Logging - רישום מלא
- *
- * 🚀 PHASE 3: LIMITED DELETE - מחיקה מוגבלת (50 items max)
- */
-exports.deleteUserDataSelective = functions.https.onCall(async (data, context) => {
-  const startTime = Date.now();
-
-  try {
-    console.log('🚀 =================================');
-    console.log('🗑️  DELETE USER DATA SELECTIVE');
-    console.log('🚀 PHASE 3: LIMITED DELETE (50 items max)');
-    console.log('🚀 =================================');
-
-    // ============================================
-    // 🔒 LAYER 1: Authentication Check
-    // ============================================
-    if (!context.auth) {
-      throw new functions.https.HttpsError('unauthenticated', 'נדרשת התחברות');
-    }
-
-    const callerEmail = context.auth.token.email;
-    const callerUid = context.auth.uid;
-
-    console.log(`👤 Caller: ${callerEmail} (${callerUid})`);
-
-    // ============================================
-    // 🔒 LAYER 2: Authorization Check (Admin Only)
-    // ============================================
-    const adminSnapshot = await db.collection('employees').where('email', '==', callerEmail).get();
-
-    if (adminSnapshot.empty) {
-      console.error(`❌ User ${callerEmail} not found in employees collection`);
-      throw new functions.https.HttpsError('permission-denied', 'אין הרשאות מנהל');
-    }
-
-    const adminData = adminSnapshot.docs[0].data();
-    if (!adminData.isAdmin && adminData.role !== 'admin') {
-      console.error(`❌ User ${callerEmail} is not admin: isAdmin=${adminData.isAdmin}, role=${adminData.role}`);
-      throw new functions.https.HttpsError('permission-denied', 'רק מנהלים יכולים למחוק נתונים');
-    }
-
-    console.log(`✅ Admin verified: ${callerEmail}`);
-
-    // ============================================
-    // 🔒 LAYER 3: Input Validation
-    // ============================================
-    const validatedData = validateDeletionRequest(data);
-    console.log(`✅ Validation passed: ${validatedData.totalItems} items to process`);
-
-    // ============================================
-    // 🔒 LAYER 4: Rate Limiting
-    // ============================================
-    if (!validatedData.dryRun) {
-      const rateLimit = await checkRateLimit(db, callerEmail);
-      console.log(`✅ Rate limit check passed: ${rateLimit.remainingInWindow} deletions remaining`);
-    }
-
-    // ============================================
-    // 🔒 LAYER 5: Suspicious Activity Check
-    // ============================================
-    const suspiciousCheck = await checkSuspiciousActivity(db, callerEmail);
-    if (suspiciousCheck.suspicious) {
-      console.warn(`⚠️ Suspicious activity detected for ${callerEmail}`);
-      // בשלב זה רק מתריעים, לא חוסמים
-    }
-
-    // ============================================
-    // 🔒 LAYER 6: Ownership Verification
-    // ============================================
-    const verifiedOwnership = await verifyAllOwnership(db, validatedData.userEmail, {
-      taskIds: validatedData.taskIds,
-      timesheetIds: validatedData.timesheetIds,
-      approvalIds: validatedData.approvalIds
-    });
-
-    console.log(`✅ Ownership verified: ${verifiedOwnership.totalVerified} items belong to ${validatedData.userEmail}`);
-
-    // ============================================
-    // 🔒 LAYER 7: Execute Deletion (or Dry Run)
-    // ============================================
-    const result = await executeDeletion(db, verifiedOwnership, validatedData.dryRun);
-
-    // ============================================
-    // 🔒 LAYER 8: Audit Logging
-    // ============================================
-    await logDeletionAttempt(db, {
-      adminEmail: callerEmail,
-      userEmail: validatedData.userEmail,
-      requestData: {
-        ...validatedData,
-        adminUid: callerUid
-      },
-      verifiedOwnership,
-      result,
-      dryRun: validatedData.dryRun,
-      success: true
-    });
-
-    const executionTime = Date.now() - startTime;
-
-    console.log('🚀 =================================');
-    console.log(`✅ SUCCESS (${executionTime}ms)`);
-    console.log(`   Mode: ${validatedData.dryRun ? 'DRY RUN' : 'REAL DELETION'}`);
-    console.log(`   Items: ${result.deletedCounts.total}`);
-    console.log('🚀 =================================');
-
-    // ============================================
-    // Response
-    // ============================================
-    return {
-      success: true,
-      dryRun: validatedData.dryRun,
-      phase: 'phase_3_limited',
-      deletionEnabled: DELETION_ENABLED,
-      message: validatedData.dryRun
-        ? `✅ Preview: ${result.deletedCounts.total} פריטים יימחקו`
-        : `✅ נמחקו ${result.deletedCounts.total} פריטים${result.deletedCounts.orphanedApprovals ? ` (כולל ${result.deletedCounts.orphanedApprovals} orphaned approvals)` : ''}`,
-      deletedCounts: result.deletedCounts,
-      preview: result.preview,
-      executionTime: `${executionTime}ms`
-    };
-
-  } catch (error) {
-    console.error('❌ Error in deleteUserDataSelective:', error);
-
-    // רישום שגיאה ב-audit log
-    try {
-      await logDeletionAttempt(db, {
-        adminEmail: context.auth?.token?.email || 'unknown',
-        userEmail: data?.userEmail || 'unknown',
-        requestData: data || {},
-        verifiedOwnership: {},
-        result: null,
-        dryRun: data?.dryRun || false,
-        success: false,
-        error
-      });
-    } catch (auditError) {
-      console.error('❌ Failed to log error to audit:', auditError);
-    }
-
-    // זריקת השגיאה הלאה
-    if (error.code && error.code.startsWith('functions/')) {
-      throw error;
-    }
-
-    throw new functions.https.HttpsError(
-      'internal',
-      error.message || 'שגיאה במחיקת נתונים'
-    );
-  }
-});
-
-// ===============================
-// Fee Agreement Functions - הסכמי שכר טרחה
-// ===============================
-
-/**
- * uploadFeeAgreement - העלאת הסכם שכר טרחה ללקוח
- * Admin only - מאובטח
- *
- * @param {Object} data
- * @param {string} data.clientId - מזהה הלקוח
- * @param {string} data.fileName - שם הקובץ המקורי
- * @param {string} data.fileData - תוכן הקובץ ב-base64
- * @param {string} data.fileType - סוג הקובץ (mime type)
- * @param {number} data.fileSize - גודל הקובץ בבייטים
- */
-exports.uploadFeeAgreement = functions.https.onCall(async (data, context) => {
-  try {
-    console.log('📄 Starting uploadFeeAgreement...');
-
-    // 1. Authorization - Admin only
-    const user = await checkUserPermissions(context);
-
-    if (user.role !== 'admin') {
-      throw new functions.https.HttpsError(
-        'permission-denied',
-        'רק מנהלים יכולים להעלות הסכמי שכר טרחה'
-      );
-    }
-
-    // 2. Input Validation
-    if (!data.clientId || typeof data.clientId !== 'string') {
-      throw new functions.https.HttpsError(
-        'invalid-argument',
-        'חסר מזהה לקוח'
-      );
-    }
-
-    if (!data.fileName || typeof data.fileName !== 'string') {
-      throw new functions.https.HttpsError(
-        'invalid-argument',
-        'חסר שם קובץ'
-      );
-    }
-
-    if (!data.fileData || typeof data.fileData !== 'string') {
-      throw new functions.https.HttpsError(
-        'invalid-argument',
-        'חסר תוכן קובץ'
-      );
-    }
-
-    if (!data.fileType || typeof data.fileType !== 'string') {
-      throw new functions.https.HttpsError(
-        'invalid-argument',
-        'חסר סוג קובץ'
-      );
-    }
-
-    // 3. File Type Validation
-    const allowedTypes = ['application/pdf', 'image/jpeg', 'image/png', 'image/webp'];
-    if (!allowedTypes.includes(data.fileType)) {
-      throw new functions.https.HttpsError(
-        'invalid-argument',
-        'סוג קובץ לא נתמך. יש להעלות PDF או תמונה (JPEG/PNG/WebP)'
-      );
-    }
-
-    // 4. File Size Validation (max 6MB for base64)
-    const maxSizeBytes = 6 * 1024 * 1024;
-    const fileSize = data.fileSize || Buffer.from(data.fileData, 'base64').length;
-
-    if (fileSize > maxSizeBytes) {
-      throw new functions.https.HttpsError(
-        'invalid-argument',
-        'הקובץ גדול מדי. גודל מקסימלי: 6MB'
-      );
-    }
-
-    // 5. Verify Client Exists
-    const clientDoc = await db.collection('clients').doc(data.clientId).get();
-
-    if (!clientDoc.exists) {
-      throw new functions.https.HttpsError(
-        'not-found',
-        'לקוח לא נמצא'
-      );
-    }
-
-    // 6. Generate unique ID and storage path
-    const agreementId = `agreement_${Date.now()}`;
-    const fileExtension = data.fileName.split('.').pop() || 'pdf';
-    const sanitizedFileName = `${agreementId}.${fileExtension}`;
-    const storagePath = `clients/${data.clientId}/agreements/${sanitizedFileName}`;
-
-    // 7. Upload to Firebase Storage
-    const bucket = admin.storage().bucket();
-    const file = bucket.file(storagePath);
-
-    // Decode base64 and upload
-    const fileBuffer = Buffer.from(data.fileData, 'base64');
-
-    await file.save(fileBuffer, {
-      metadata: {
-        contentType: data.fileType,
-        metadata: {
-          uploadedBy: user.email,
-          originalName: data.fileName,
-          clientId: data.clientId
-        }
-      }
-    });
-
-    // 8. Get download URL
-    await file.makePublic();
-    const downloadUrl = `https://storage.googleapis.com/${bucket.name}/${storagePath}`;
-
-    // Alternative: Use signed URL (more secure but expires)
-    // const [signedUrl] = await file.getSignedUrl({
-    //   action: 'read',
-    //   expires: Date.now() + 365 * 24 * 60 * 60 * 1000 // 1 year
-    // });
-
-    // 9. Create agreement metadata
-    const agreementData = {
-      id: agreementId,
-      fileName: sanitizedFileName,
-      originalName: data.fileName,
-      storagePath: storagePath,
-      downloadUrl: downloadUrl,
-      fileType: data.fileType,
-      fileSize: fileSize,
-      uploadedAt: admin.firestore.Timestamp.now(),
-      uploadedBy: user.email
-    };
-
-    // 10. Update client document with new agreement
-    const clientData = clientDoc.data();
-    const existingAgreements = clientData.feeAgreements || [];
-
-    await db.collection('clients').doc(data.clientId).update({
-      feeAgreements: [...existingAgreements, agreementData],
-      lastModifiedBy: user.username,
-      lastModifiedAt: admin.firestore.FieldValue.serverTimestamp()
-    });
-
-    // 11. Audit Log
-    await logAction('UPLOAD_FEE_AGREEMENT', user.uid, user.username, {
-      clientId: data.clientId,
-      clientName: clientData.fullName || clientData.clientName,
-      agreementId: agreementId,
-      fileName: data.fileName,
-      fileType: data.fileType,
-      fileSize: fileSize
-    });
-
-    console.log(`✅ Fee agreement uploaded successfully: ${agreementId} for client ${data.clientId}`);
-
-    return {
-      success: true,
-      agreement: {
-        ...agreementData,
-        uploadedAt: new Date().toISOString()
-      },
-      message: 'הסכם שכר טרחה הועלה בהצלחה'
-    };
-
-  } catch (error) {
-    console.error('Error in uploadFeeAgreement:', error);
-
-    if (error instanceof functions.https.HttpsError) {
-      throw error;
-    }
-
-    throw new functions.https.HttpsError(
-      'internal',
-      `שגיאה בהעלאת הסכם: ${error.message}`
-    );
-  }
-});
-
-/**
- * deleteFeeAgreement - מחיקת הסכם שכר טרחה
- * Admin only - מאובטח
- *
- * @param {Object} data
- * @param {string} data.clientId - מזהה הלקוח
- * @param {string} data.agreementId - מזהה ההסכם למחיקה
- */
-exports.deleteFeeAgreement = functions.https.onCall(async (data, context) => {
-  try {
-    console.log('🗑️ Starting deleteFeeAgreement...');
-
-    // 1. Authorization - Admin only
-    const user = await checkUserPermissions(context);
-
-    if (user.role !== 'admin') {
-      throw new functions.https.HttpsError(
-        'permission-denied',
-        'רק מנהלים יכולים למחוק הסכמי שכר טרחה'
-      );
-    }
-
-    // 2. Input Validation
-    if (!data.clientId || typeof data.clientId !== 'string') {
-      throw new functions.https.HttpsError(
-        'invalid-argument',
-        'חסר מזהה לקוח'
-      );
-    }
-
-    if (!data.agreementId || typeof data.agreementId !== 'string') {
-      throw new functions.https.HttpsError(
-        'invalid-argument',
-        'חסר מזהה הסכם'
-      );
-    }
-
-    // 3. Verify Client Exists
-    const clientDoc = await db.collection('clients').doc(data.clientId).get();
-
-    if (!clientDoc.exists) {
-      throw new functions.https.HttpsError(
-        'not-found',
-        'לקוח לא נמצא'
-      );
-    }
-
-    const clientData = clientDoc.data();
-    const existingAgreements = clientData.feeAgreements || [];
-
-    // 4. Find the agreement to delete
-    const agreementToDelete = existingAgreements.find(a => a.id === data.agreementId);
-
-    if (!agreementToDelete) {
-      throw new functions.https.HttpsError(
-        'not-found',
-        'הסכם לא נמצא'
-      );
-    }
-
-    // 5. Delete from Firebase Storage
-    try {
-      const bucket = admin.storage().bucket();
-      const file = bucket.file(agreementToDelete.storagePath);
-      await file.delete();
-      console.log(`✅ Deleted file from storage: ${agreementToDelete.storagePath}`);
-    } catch (storageError) {
-      // Log but don't fail if storage deletion fails (file might not exist)
-      console.warn(`⚠️ Could not delete file from storage: ${storageError.message}`);
-    }
-
-    // 6. Remove from Firestore
-    const updatedAgreements = existingAgreements.filter(a => a.id !== data.agreementId);
-
-    await db.collection('clients').doc(data.clientId).update({
-      feeAgreements: updatedAgreements,
-      lastModifiedBy: user.username,
-      lastModifiedAt: admin.firestore.FieldValue.serverTimestamp()
-    });
-
-    // 7. Audit Log
-    await logAction('DELETE_FEE_AGREEMENT', user.uid, user.username, {
-      clientId: data.clientId,
-      clientName: clientData.fullName || clientData.clientName,
-      agreementId: data.agreementId,
-      fileName: agreementToDelete.originalName || agreementToDelete.fileName
-    });
-
-    console.log(`✅ Fee agreement deleted successfully: ${data.agreementId} from client ${data.clientId}`);
-
-    return {
-      success: true,
-      message: 'הסכם שכר טרחה נמחק בהצלחה'
-    };
-
-  } catch (error) {
-    console.error('Error in deleteFeeAgreement:', error);
-
-    if (error instanceof functions.https.HttpsError) {
-      throw error;
-    }
-
-    throw new functions.https.HttpsError(
-      'internal',
-      `שגיאה במחיקת הסכם: ${error.message}`
-    );
-  }
-});
 
 // ═══════════════════════════════════════════════════════════════
 // 📊 Workload Analytics Functions - Performance Optimized
@@ -7540,125 +5531,38 @@ exports.deleteFeeAgreement = functions.https.onCall(async (data, context) => {
 const { getTeamWorkloadData } = require('./workload-analytics');
 exports.getTeamWorkloadData = getTeamWorkloadData;
 
-// ═══════════════════════════════════════════════════════════════
-// 🔍 Daily Invariant Check - Data Integrity Monitor
-// ═══════════════════════════════════════════════════════════════
+// Fee Agreements Functions (imported from ./fee-agreements)
+const feeAgreements = require('./fee-agreements');
+exports.uploadFeeAgreement = feeAgreements.uploadFeeAgreement;
+exports.deleteFeeAgreement = feeAgreements.deleteFeeAgreement;
 
-// TODO: כשישודרג Twilio — להוסיף שליחת SMS בפער
-// מספר יעד: +972549539238
+// Scheduled Functions (imported from ./scheduled)
+const scheduled = require('./scheduled');
+exports.dailyTaskReminders = scheduled.dailyTaskReminders;
+exports.dailyBudgetWarnings = scheduled.dailyBudgetWarnings;
+exports.dailyInvariantCheck = scheduled.dailyInvariantCheck;
 
-exports.dailyInvariantCheck = onSchedule({
-  schedule: '0 6 * * *',
-  timeZone: 'Asia/Jerusalem',
-  region: 'us-central1'
-}, async () => {
-  const SKIP_CLIENTS = ['2025003'];
-  const TOLERANCE = 0.02;
-  const discrepancies = [];
+// WhatsApp Functions (imported from ./whatsapp)
+const whatsapp = require('./whatsapp');
+exports.sendBroadcastMessage = whatsapp.sendBroadcastMessage;
+exports.sendWhatsAppApprovalNotification = whatsapp.sendWhatsAppApprovalNotification;
+exports.whatsappWebhook = whatsapp.whatsappWebhook;
+exports.onApprovalCreated = whatsapp.onApprovalCreated;
 
-  try {
-    console.log('🔍 Starting daily invariant check...');
+// Metrics Functions (imported from ./metrics)
+const metrics = require('./metrics');
+exports.getUserMetrics = metrics.getUserMetrics;
+exports.updateMetricsOnTaskChange = metrics.updateMetricsOnTaskChange;
 
-    const clientsSnapshot = await db.collection('clients').get();
-    console.log(`📊 Checking ${clientsSnapshot.size} clients`);
+// Admin Functions (imported from ./admin)
+const adminOps = require('./admin');
+exports.createUser = adminOps.createUser;
+exports.updateUser = adminOps.updateUser;
+exports.blockUser = adminOps.blockUser;
+exports.deleteUser = adminOps.deleteUser;
+exports.getUserFullDetails = adminOps.getUserFullDetails;
+exports.logActivity = adminOps.logActivity;
+exports.deleteUserData = adminOps.deleteUserData;
+exports.deleteUserDataSelective = adminOps.deleteUserDataSelective;
 
-    for (const clientDoc of clientsSnapshot.docs) {
-      const clientId = clientDoc.id;
-
-      if (SKIP_CLIENTS.includes(clientId)) {
-        continue;
-      }
-
-      try {
-        const clientData = clientDoc.data();
-        const clientName = clientData.clientName || clientData.name || clientId;
-        const services = clientData.services || [];
-
-        if (services.length === 0) {
-          continue;
-        }
-
-        // Read all timesheet entries for this client
-        const timesheetSnapshot = await db.collection('timesheet_entries')
-          .where('clientId', '==', clientId)
-          .get();
-
-        // Group minutes by serviceId
-        const serviceMinutes = {};
-        timesheetSnapshot.forEach(doc => {
-          const entry = doc.data();
-          const serviceId = entry.serviceId;
-          if (serviceId) {
-            serviceMinutes[serviceId] = (serviceMinutes[serviceId] || 0) + (entry.minutes || 0);
-          }
-        });
-
-        // Check each service
-        for (const service of services) {
-          const serviceId = service.id;
-          if (!serviceId) continue;
-
-          const cardHoursUsed = service.hoursUsed || 0;
-          const timesheetMinutes = serviceMinutes[serviceId] || 0;
-          const timesheetHoursUsed = timesheetMinutes / 60;
-          const gap = Math.abs(cardHoursUsed - timesheetHoursUsed);
-
-          if (gap > TOLERANCE) {
-            discrepancies.push({
-              clientId,
-              clientName,
-              serviceId,
-              serviceName: service.name || service.type || serviceId,
-              cardHoursUsed: parseFloat(cardHoursUsed.toFixed(2)),
-              timesheetHoursUsed: parseFloat(timesheetHoursUsed.toFixed(2)),
-              gap: parseFloat(gap.toFixed(2))
-            });
-          }
-        }
-      } catch (clientError) {
-        console.error(`⚠️ Error checking client ${clientId}:`, clientError.message);
-        // Continue to next client
-      }
-    }
-
-    // Save result to system_health_checks
-    if (discrepancies.length > 0) {
-      await db.collection('system_health_checks').add({
-        type: 'invariant_check',
-        status: 'FAIL',
-        timestamp: admin.firestore.FieldValue.serverTimestamp(),
-        discrepanciesCount: discrepancies.length,
-        discrepancies,
-        message: `נמצאו ${discrepancies.length} פערים בנתוני שעות`
-      });
-      console.log(`❌ Invariant check FAILED — ${discrepancies.length} discrepancies found`);
-    } else {
-      await db.collection('system_health_checks').add({
-        type: 'invariant_check',
-        status: 'PASS',
-        timestamp: admin.firestore.FieldValue.serverTimestamp(),
-        discrepanciesCount: 0,
-        discrepancies: [],
-        message: 'כל הנתונים תקינים'
-      });
-      console.log('✅ Invariant check PASSED — no discrepancies');
-    }
-
-  } catch (error) {
-    console.error('❌ Invariant check ERROR:', error);
-    try {
-      await db.collection('system_health_checks').add({
-        type: 'invariant_check',
-        status: 'ERROR',
-        timestamp: admin.firestore.FieldValue.serverTimestamp(),
-        discrepanciesCount: 0,
-        discrepancies: [],
-        message: `שגיאה בבדיקת תקינות: ${error.message}`
-      });
-    } catch (saveError) {
-      console.error('❌ Failed to save error status:', saveError);
-    }
-  }
-});
-
-console.log('✅ Law Office Functions loaded successfully (including 10 Master Admin functions + User Metrics + setAdminClaims + Task Approval System + WhatsApp Broadcast + WhatsApp Smart Bot 🤖 + Delete User Data + Delete User Data Selective 🔒 + Fee Agreements 📄 + Workload Analytics 📊 + Daily Invariant Check 🔍)');
+console.log('✅ Law Office Functions loaded successfully');
